Move AccessControlInfo to presto-spi
AccessControlInfo is an inner class in the Analysis. This class will
be used to store accessControl check across all analyzers, hence
moving this to presto-spi.
